requirements.txt
requests
python-dateutil
PyYAML
tqdm